requirements.txt
praw==7.7.1
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
transformers>=4.37.0
torch>=2.2.0
streamlit>=1.31.0
plotly>=5.18.0
python-dotenv>=1.0.0
wordcloud>=1.9.3
matplotlib>=3.8.0 